1.3

+++ Telegraph/TGChannelManagementSignals.h
@@ -72,7 +72,7 @@ typedef struct {
 + (SSignal *)channelBannedMembers:(int64_t)peerId accessHash:(int64_t)accessHash offset:(NSUInteger)offset count:(NSUInteger)count;
 + (SSignal *)channelMembers:(int64_t)peerId accessHash:(int64_t)accessHash offset:(NSUInteger)offset count:(NSUInteger)count;
 + (SSignal *)channelAdmins:(int64_t)peerId accessHash:(int64_t)accessHash offset:(NSUInteger)offset count:(NSUInteger)count;
-+ (SSignal *)channelAdmins:(int64_t)peerId accessHash:(int64_t)accessHash offset:(NSUInteger)offset count:(NSUInteger)count hash:(int32_t)hash;
++ (SSignal *)channelAdmins:(int64_t)peerId accessHash:(int64_t)accessHash offset:(NSUInteger)offset count:(NSUInteger)count hash:(int64_t)hash;
 + (SSignal *)channelInviterUser:(int64_t)peerId accessHash:(int64_t)accessHash;
 
 + (SSignal *)deleteChannel:(int64_t)peerId accessHash:(int64_t)accessHash;
